requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
distro==1.9.0
faiss-cpu==1.12.0
fastapi==0.119.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.0
joblib==1.5.2
motor==3.6.0
nltk==3.9.2
numpy==2.3.3
openai==2.3.0
packaging==25.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic==2.12.0
pydantic_core==2.41.1
PyPDF2==3.0.1
PyJWT==2.9.0
python-dotenv==1.1.1
python-multipart==0.0.20
regex==2025.9.18
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
